chore: rename internal as common
