+++ ghx/idgen/idgen.go
@@ -4,8 +4,8 @@ import (
 	"path/filepath"
 	"strconv"
 
+	"github.com/aweris/gale/common/fs"
 	"github.com/aweris/gale/ghx/context"
-	"github.com/aweris/gale/internal/fs"
 )
 
 const (
