+++ ghx/step_docker.go
@@ -6,10 +6,10 @@ import (
 
 	"dagger.io/dagger"
 
+	"github.com/aweris/gale/common/log"
 	"github.com/aweris/gale/ghx/context"
 	"github.com/aweris/gale/ghx/core"
 	"github.com/aweris/gale/ghx/task"
-	"github.com/aweris/gale/internal/log"
 )
 
 var (
