+++ ghx/workflow_run.go
@@ -4,11 +4,11 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/aweris/gale/common/log"
 	"github.com/aweris/gale/ghx/context"
 	"github.com/aweris/gale/ghx/core"
 	"github.com/aweris/gale/ghx/idgen"
 	"github.com/aweris/gale/ghx/task"
-	"github.com/aweris/gale/internal/log"
 )
 
 // planWorkflow plans the workflow and returns the workflow runner.
