+++ go.work
@@ -2,5 +2,5 @@ go 1.21
 
 use (
 	.
-	./internal/mage
+	mage
 )
