--- internal/mage/magefile.go
@@ -1,11 +1,0 @@
-//go:build mage
-
-package main
-
-import (
-	//mage:import docker
-	_ "github.com/aweris/gale/internal/mage/docker"
-
-	//mage:import dev
-	_ "github.com/aweris/gale/internal/mage/dev"
-)
