+++ mage/.gitignore
